requirements.txt
fastapi
pydantic
uvicorn
pyyaml
